Move _lstm_fn from recurrent to recurrent_internals
